requirements.txt
selenium
pillow
pytesseract
python-telegram-bot==13.15
